requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
PyMySQL>=1.1.0
Werkzeug>=3.0.0
fpdf2>=2.7.0
pywebview>=4.4
